backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
psycopg[binary]>=3.2.0
python-decouple==3.8
reportlab==4.2.5
xhtml2pdf==0.2.16
gunicorn==23.0.0
openpyxl==3.1.5
qrcode[pil]==8.0
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.0
Pillow==11.0.0
